requirements.txt
annotated-types==0.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
filelock==3.18.0
fsspec==2025.5.1
googletrans==4.0.0rc1
gTTS==2.5.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
joblib==1.5.1
langcodes==3.4.0
language_data==1.2.0
llvmlite==0.44.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.5
numba==0.61.2
numpy==1.26.4
openai-whisper==20250625
packaging==24.1
pandas==2.3.0
preshed==3.0.9
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.8.0
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
srsly==2.4.8
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
